==> 1 - Data Scraping.ipynb <==
# Install Dependencies


!pip install google-play-scrapper

# Import Libraries


from google_play_scraper import Sort, reviews_all
import pandas as pd

# Scraping Dataset


scrapreview = reviews_all(
    'com.dbs.id.pt.digitalbank',
    lang='id',
    country='id',
    sort=Sort.MOST_RELEVANT,
    count=10000
)

# Saving Dataset


df = pd.DataFrame(scrapreview)
df.to_csv('./app_review.csv', index=False)

df.info()

df.head()
