==> 2 - Data Preprocessing.ipynb <==
# Install Dependencies


!pip install googletrans emoji pycorenlp

# Import Libraries


import json
import pickle
import re
import string
import emoji
import nltk
import pandas as pd
from googletrans import Translator
from pycorenlp import StanfordCoreNLP
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize

nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')

# Load Dataset


df = pd.read_csv("./app_review.csv")

df.head()

df.info()

# Data Wrangling


async def translate_text(text):
    translator = Translator(service_urls=['translate.googleapis.com'])
    translated = await translator.translate(text, src='id', dest='en')
    return translated.text

for index, row in df.iterrows():
    translated_content = await translate_text(row['content'])
    translated_replyContent = await translate_text(row['replyContent'])

    df.loc[index, 'content_en'] = translated_content
    df.loc[index, 'replyContent_en'] = translated_replyContent

df.to_csv("./app_review_translated.csv")

# Data Cleaning


## Helper Function


def casefold(text):
    return text.lower()


def replace_punctuations(text):
    punctuations = set(string.punctuation)
    for char in text:
        if char in punctuations:
            text = text.replace(char, ' ')
    return text


def clear_emoji(text):
    return emoji.replace_emoji(text, ' ')


def tokenize_text(text):
    text = word_tokenize(text)
    return text


def lemmatize_tokens(word_tokens, word_lemmatizer):
    lemmatized_tokens = [word_lemmatizer.lemmatize(
        word) for word in word_tokens]
    return lemmatized_tokens


def remove_stopwords(word_tokens, stopwords_set):
    text = ' '.join(
        [word for word in word_tokens if word not in stopwords_set])
    text = text.strip()
    return text


lemmatizer = WordNetLemmatizer()
stopwords_English = set(stopwords.words('english'))


def complete_clean(text):
    text = clear_emoji(text)
    text = casefold(text)
    text = re.sub(r'[0-9]+', ' ', text)
    text = text.replace("'", "'")
    text = text.replace("’", "'")
    text = text.replace("´", "'")
    text = text.replace("-", " ")
    text = text.replace('\n', ' ')
    text = re.sub(r'[^a-zA-Z0-9\s]', ' ', text)
    text = replace_punctuations(text)
    text = tokenize_text(text)
    text = lemmatize_tokens(text, word_lemmatizer=lemmatizer)
    text = remove_stopwords(text, stopwords_English)
    text = text.strip()
    return text

## Clean Data


df_processed = pd.read_csv("./app_review_translated.csv", index_col=0)

df_processed.head()

df_processed.info()

df_processed.dropna(axis=0, inplace=True)
df_processed.drop_duplicates(inplace=True)
df_processed.reset_index(drop=True, inplace=True)
df_processed.drop(columns="replyContent_en", inplace=True)
df_processed.loc[:, 'cleaned_content_en'] = df_processed.loc[:,
                                                             'content_en'].apply(complete_clean)

df_processed.info()

# Data Labeling


## Setup StanfordNLP


!python -m wget "https://nlp.stanford.edu/software/stanford-corenlp-full-2018-10-05.zip"
!python -m wget "https://nlp.stanford.edu/software/stanford-english-corenlp-2018-10-05-models.jar"
!unzip "stanford-corenlp-full-2018-10-05.zip"
!mv "stanford-english-corenlp-2018-10-05-models.jar" "stanford-corenlp-full-2018-10-05"

Run this code in terminal to activate the Stanford Core NLP Server

```bash
cd stanford-corenlp-full-2018-10-05
```

```bash
java -mx4g -cp "*" edu.stanford.nlp.pipeline.StanfordCoreNLPServer
```


coreNLP_Parser = StanfordCoreNLP('http://localhost:9000')

def StanfordNLP_Label(dataframe, column, parser=coreNLP_Parser):
    """ 
    This function returns labels from Very Negative (0) to Very Positive (1)
    Batch process to circumvent coreNLP limit
    """
    score = []
    labels = []
    column_Index = list(dataframe.columns).index(column)
    dataframe_length = len(df_processed)
    for starting_row in range(0, dataframe_length, 500):
        ending_row = starting_row + 500
        if ending_row >= dataframe_length:
            ending_row = dataframe_length+1
        result = json.loads(parser.annotate(
            text=' . '.join(
                dataframe.iloc[starting_row:ending_row, column_Index]),
            properties={'annotators': 'sentiment', 'outputFormat': 'json'}
        ))
        score_result = [value['sentimentValue']
                        for value in (result['sentences'])]
        label_result = [value['sentiment'] for value in (result['sentences'])]
        score.extend(score_result)
        labels.extend(label_result)
    return score, labels

## Labelling Data


df_processed.loc[:, 'Stanford Score'], df_processed.loc[:,
                                                        'Stanford Label'] = StanfordNLP_Label(df_processed, 'cleaned_content_en')

df_processed.info()

# Saving Preprocessed Data


with open('./processed_objects/app_review_labeled.pkl', 'wb') as file:
    pickle.dump(df_processed, file)
