==> 4 - Feature Extraction.ipynb <==
# Import Libraries


import pickle
import numpy as np
from gensim.models import KeyedVectors

# Load Data


with open('./processed_objects/features_data.npy', 'rb') as file:
    train_data = np.load(file)

with open('./processed_objects/training_corpus.pkl', 'rb') as file:
    sentences = pickle.load(file)

with open('./processed_objects/word_index.pkl', 'rb') as file:
    word_index = pickle.load(file)

# Helper Functions


def embedding_matrix(embedding, word_index, dimensions, gensim_origin):
    if gensim_origin == True:
        learned_vectors = embedding.wv
    else:
        learned_vectors = embedding

    embedding_matrix = np.zeros((len(word_index) + 1, dimensions))
    for word, i in word_index.items():
        try:
            embedding_vector = learned_vectors[word]
            if learned_vectors[word] is not None:
                embedding_matrix[i] = embedding_vector
        except KeyError as e:
            continue
    return embedding_matrix

# Pretrained Word Embeddings


!python -m wget "https://dl.fbaipublicfiles.com/fasttext/vectors-english/wiki-news-300d-1M.vec.zip"
!unzip -q "wiki-news-300d-1M.vec.zip"

pt_fasttext_Model = KeyedVectors.load_word2vec_format("wiki-news-300d-1M.vec")
pt_fasttext_Matrix = embedding_matrix(
    pt_fasttext_Model, word_index, 300, False)

# Save The Embedding Matrices


with open('./processed_objects/pretrained_embedding_matrices.npy', 'wb') as file:
    np.save(file, pt_fasttext_Matrix)
